@@ -13,8 +13,6 @@ declare(strict_types=1);
 
 namespace Jgut\Slim\Exception\Whoops\Renderer;
 
-use Jgut\Slim\Exception\Whoops\Inspector;
-use Whoops\Handler\Handler;
 use Whoops\Handler\PrettyPageHandler;
 
 /**
@@ -24,22 +22,11 @@ class HtmlRenderer extends PrettyPageHandler
 {
     use RendererTrait;
 
-    public function __construct(string $defaultTitle = 'Slim Application error')
-    {
+    public function __construct(
+        protected string $defaultTitle = 'Slim Application error',
+    ) {
         parent::__construct();
 
-        $this->defaultTitle = $defaultTitle;
         $this->setPageTitle($defaultTitle);
-    }
-
-    /**
-     * @inheritDoc
-     */
-    public function handle()
-    {
-        $exception = $this->getException();
-        $this->setInspector(new Inspector($exception));
-
-        return parent::handle() ?? Handler::QUIT;
     }
 }
