@@ -18,6 +18,7 @@ use PHPUnit\Framework\TestCase;
 use Psr\Http\Message\ServerRequestInterface;
 use Slim\Exception\HttpForbiddenException;
 use Whoops\Exception\Inspector;
+use Whoops\Run as Whoops;
 
 /**
  * @internal
@@ -31,22 +32,27 @@ class XmlRendererTest extends TestCase
         $inspector = new Inspector($exception);
 
         $renderer = new XmlRenderer();
-        $renderer->addTraceToOutput(true);
         $renderer->setException($exception);
         $renderer->setInspector($inspector);
+        $renderer->setRun(new Whoops());
 
         ob_start();
         $renderer->handle();
         $output = ob_get_clean();
 
-        $expected = <<<'EXPECTED'
-        <?xml version="1.0" encoding="UTF-8" standalone="yes"?>
-        <error>
-          <message><![CDATA[403 Forbidden]]></message>
-          <type>Slim\Exception\HttpForbiddenException</type>
-          <trace>
+        $file = __FILE__;
 
-        EXPECTED;
+        $expected /** @lang xml */
+            = <<<EXPECTED
+            <?xml version="1.0" encoding="UTF-8" standalone="yes"?>
+            <error>
+              <type>Slim\\Exception\\HttpForbiddenException</type>
+              <message><![CDATA[403 Forbidden]]></message>
+              <code>403</code>
+              <file>{$file}</file>
+              <line>31</line>
+              <trace>
+            EXPECTED;
         static::assertStringContainsString($expected, $output);
     }
 
@@ -57,19 +63,23 @@ class XmlRendererTest extends TestCase
         $inspector = new Inspector($exception);
 
         $renderer = new XmlRenderer();
-        $renderer->addTraceToOutput(true);
         $renderer->setException($exception);
         $renderer->setInspector($inspector);
         $renderer->setPrettify(false);
+        $renderer->setRun(new Whoops());
 
         ob_start();
         $renderer->handle();
         $output = ob_get_clean();
 
-        $expected = <<<'EXPECTED'
-        <?xml version="1.0" encoding="UTF-8" standalone="yes"?>
-        <error><message><![CDATA[403 Forbidden]]></message><type>Slim\Exception\HttpForbiddenException</type><trace>
-        EXPECTED;
+        $expected = '<?xml version="1.0" encoding="UTF-8" standalone="yes"?>' . "\n"
+            . '<error>'
+            . '<type>Slim\Exception\HttpForbiddenException</type>'
+            . '<message><![CDATA[403 Forbidden]]></message>'
+            . '<code>403</code>'
+            . '<file>' . __FILE__ . '</file>'
+            . '<line>62</line>'
+            . '<trace>';
         static::assertStringContainsString($expected, $output);
     }
 
@@ -83,18 +93,26 @@ class XmlRendererTest extends TestCase
         $renderer->addTraceToOutput(false);
         $renderer->setException($exception);
         $renderer->setInspector($inspector);
+        $renderer->setRun(new Whoops());
 
         ob_start();
         $renderer->handle();
         $output = ob_get_clean();
 
-        $expected = <<<'EXPECTED'
-        <?xml version="1.0" encoding="UTF-8" standalone="yes"?>
-        <error>
-          <message><![CDATA[403 Forbidden]]></message>
-        </error>
+        $file = __FILE__;
 
-        EXPECTED;
+        $expected /** @lang xml */
+            = <<<EXPECTED
+            <?xml version="1.0" encoding="UTF-8" standalone="yes"?>
+            <error>
+              <type>Slim\\Exception\\HttpForbiddenException</type>
+              <message><![CDATA[403 Forbidden]]></message>
+              <code>403</code>
+              <file>{$file}</file>
+              <line>89</line>
+            </error>
+
+            EXPECTED;
         static::assertEquals($expected, $output);
     }
 
@@ -109,16 +127,20 @@ class XmlRendererTest extends TestCase
         $renderer->setException($exception);
         $renderer->setInspector($inspector);
         $renderer->setPrettify(false);
+        $renderer->setRun(new Whoops());
 
         ob_start();
         $renderer->handle();
         $output = ob_get_clean();
 
-        $expected = <<<'EXPECTED'
-        <?xml version="1.0" encoding="UTF-8" standalone="yes"?>
-        <error><message><![CDATA[403 Forbidden]]></message></error>
-
-        EXPECTED;
+        $expected = '<?xml version="1.0" encoding="UTF-8" standalone="yes"?>' . "\n"
+            . '<error>'
+            . '<type>Slim\Exception\HttpForbiddenException</type>'
+            . '<message><![CDATA[403 Forbidden]]></message>'
+            . '<code>403</code>'
+            . '<file>' . __FILE__ . '</file>'
+            . '<line>122</line>'
+            . '</error>' . "\n";
         static::assertEquals($expected, $output);
     }
 }
